Renaming the new special header
Regarding the argument collecting, I'm thinking about a sort of communication struct, similar to what `GLOBAL` is right now
idk
