+++ ArbAsmlibsup/arbnum_global.h
